Type Alias ttf_parser::apple_layout::StateEntry

pub type StateEntry = GenericStateEntry<()>;

A legacy state entry used by StateTable.

Aliased Type

struct StateEntry {
    pub new_state: u16,
    pub flags: u16,
    pub extra: (),
}

Fields
new_state: u16

A new state.

flags: u16

Entry flags.

extra: ()

Additional data.

Use () if no data expected.

Implementations

impl<T: FromData> GenericStateEntry<T>

pub fn has_offset(&self) -> bool

Checks that entry has an offset.

pub fn value_offset(&self) -> ValueOffset

Returns a value offset.

Used by kern::format1 subtable.

pub fn has_reset(&self) -> bool

If set, reset the kerning data (clear the stack).

pub fn has_advance(&self) -> bool

If set, advance to the next glyph before going to the new state.

pub fn has_push(&self) -> bool

If set, push this glyph on the kerning stack.

pub fn has_mark(&self) -> bool

If set, remember this glyph as the marked glyph.

Used by kerx::format4 subtable.

Yes, the same as has_push.