ttf_parser::colr

Struct SweepGradient

pub struct SweepGradient<'a> {
    pub center_x: f32,
    pub center_y: f32,
    pub start_angle: f32,
    pub end_angle: f32,
    pub extend: GradientExtend,
    /* private fields */
}

A sweep gradient

Fields
center_x: f32

The x of the center.

center_y: f32

The y of the center.

start_angle: f32

The start angle.

end_angle: f32

The end angle.

extend: GradientExtend

The extend.

Implementations

impl<'a> SweepGradient<'a>

pub fn stops<'b>( &'b self, palette: u16, coords: &'a [NormalizedCoordinate], ) -> GradientStopsIter<'a, 'b>

Returns an iterator over the stops of the sweep gradient. Stops need to be sorted manually by the caller.