Struct ttf_parser::math::Constants

pub struct Constants<'a> { /* private fields */ }

A Math Constants Table.

Implementations

impl<'a> Constants<'a>

pub fn script_percent_scale_down(&self) -> i16

Percentage of scaling down for level 1 superscripts and subscripts.

pub fn script_script_percent_scale_down(&self) -> i16

Percentage of scaling down for level 2 (scriptScript) superscripts and subscripts.

pub fn delimited_sub_formula_min_height(&self) -> u16

Minimum height required for a delimited expression (contained within parentheses, etc.) to be treated as a sub-formula.

pub fn display_operator_min_height(&self) -> u16

Minimum height of n-ary operators (such as integral and summation) for formulas in display mode (that is, appearing as standalone page elements, not embedded inline within text).

pub fn math_leading(&self) -> MathValue<'a>

White space to be left between math formulas to ensure proper line spacing.

pub fn axis_height(&self) -> MathValue<'a>

Axis height of the font.

pub fn accent_base_height(&self) -> MathValue<'a>

Maximum (ink) height of accent base that does not require raising the accents.

pub fn flattened_accent_base_height(&self) -> MathValue<'a>

Maximum (ink) height of accent base that does not require flattening the accents.

pub fn subscript_shift_down(&self) -> MathValue<'a>

The standard shift down applied to subscript elements.

pub fn subscript_top_max(&self) -> MathValue<'a>

Maximum allowed height of the (ink) top of subscripts that does not require moving subscripts further down.

pub fn subscript_baseline_drop_min(&self) -> MathValue<'a>

Minimum allowed drop of the baseline of subscripts relative to the (ink) bottom of the base.

pub fn superscript_shift_up(&self) -> MathValue<'a>

Standard shift up applied to superscript elements.

pub fn superscript_shift_up_cramped(&self) -> MathValue<'a>

Standard shift of superscripts relative to the base, in cramped style.

pub fn superscript_bottom_min(&self) -> MathValue<'a>

Minimum allowed height of the (ink) bottom of superscripts that does not require moving subscripts further up.

pub fn superscript_baseline_drop_max(&self) -> MathValue<'a>

Maximum allowed drop of the baseline of superscripts relative to the (ink) top of the base.

pub fn sub_superscript_gap_min(&self) -> MathValue<'a>

Minimum gap between the superscript and subscript ink.

pub fn superscript_bottom_max_with_subscript(&self) -> MathValue<'a>

The maximum level to which the (ink) bottom of superscript can be pushed to increase the gap between superscript and subscript, before subscript starts being moved down.

pub fn space_after_script(&self) -> MathValue<'a>

Extra white space to be added after each subscript and superscript.

pub fn upper_limit_gap_min(&self) -> MathValue<'a>

Minimum gap between the (ink) bottom of the upper limit, and the (ink) top of the base operator.

pub fn upper_limit_baseline_rise_min(&self) -> MathValue<'a>

Minimum distance between baseline of upper limit and (ink) top of the base operator.

pub fn lower_limit_gap_min(&self) -> MathValue<'a>

Minimum gap between (ink) top of the lower limit, and (ink) bottom of the base operator.

pub fn lower_limit_baseline_drop_min(&self) -> MathValue<'a>

Minimum distance between baseline of the lower limit and (ink) bottom of the base operator.

pub fn stack_top_shift_up(&self) -> MathValue<'a>

Standard shift up applied to the top element of a stack.

pub fn stack_top_display_style_shift_up(&self) -> MathValue<'a>

Standard shift up applied to the top element of a stack in display style.

pub fn stack_bottom_shift_down(&self) -> MathValue<'a>

Standard shift down applied to the bottom element of a stack.

pub fn stack_bottom_display_style_shift_down(&self) -> MathValue<'a>

Standard shift down applied to the bottom element of a stack in display style.

pub fn stack_gap_min(&self) -> MathValue<'a>

Minimum gap between (ink) bottom of the top element of a stack, and the (ink) top of the bottom element.

pub fn stack_display_style_gap_min(&self) -> MathValue<'a>

Minimum gap between (ink) bottom of the top element of a stack, and the (ink) top of the bottom element in display style.

pub fn stretch_stack_top_shift_up(&self) -> MathValue<'a>

Standard shift up applied to the top element of the stretch stack.

pub fn stretch_stack_bottom_shift_down(&self) -> MathValue<'a>

Standard shift down applied to the bottom element of the stretch stack.

pub fn stretch_stack_gap_above_min(&self) -> MathValue<'a>

Minimum gap between the ink of the stretched element, and the (ink) bottom of the element above.

pub fn stretch_stack_gap_below_min(&self) -> MathValue<'a>

Minimum gap between the ink of the stretched element, and the (ink) top of the element below.

pub fn fraction_numerator_shift_up(&self) -> MathValue<'a>

Standard shift up applied to the numerator.

pub fn fraction_numerator_display_style_shift_up(&self) -> MathValue<'a>

Standard shift up applied to the numerator in display style.

pub fn fraction_denominator_shift_down(&self) -> MathValue<'a>

Standard shift down applied to the denominator.

pub fn fraction_denominator_display_style_shift_down(&self) -> MathValue<'a>

Standard shift down applied to the denominator in display style.

pub fn fraction_numerator_gap_min(&self) -> MathValue<'a>

Minimum tolerated gap between the (ink) bottom of the numerator and the ink of the fraction bar.

pub fn fraction_num_display_style_gap_min(&self) -> MathValue<'a>

Minimum tolerated gap between the (ink) bottom of the numerator and the ink of the fraction bar in display style.

pub fn fraction_rule_thickness(&self) -> MathValue<'a>

Thickness of the fraction bar.

pub fn fraction_denominator_gap_min(&self) -> MathValue<'a>

Minimum tolerated gap between the (ink) top of the denominator and the ink of the fraction bar.

pub fn fraction_denom_display_style_gap_min(&self) -> MathValue<'a>

Minimum tolerated gap between the (ink) top of the denominator and the ink of the fraction bar in display style.

pub fn skewed_fraction_horizontal_gap(&self) -> MathValue<'a>

Horizontal distance between the top and bottom elements of a skewed fraction.

pub fn skewed_fraction_vertical_gap(&self) -> MathValue<'a>

Vertical distance between the ink of the top and bottom elements of a skewed fraction.

pub fn overbar_vertical_gap(&self) -> MathValue<'a>

Distance between the overbar and the (ink) top of he base.

pub fn overbar_rule_thickness(&self) -> MathValue<'a>

Thickness of overbar.

pub fn overbar_extra_ascender(&self) -> MathValue<'a>

Extra white space reserved above the overbar.

pub fn underbar_vertical_gap(&self) -> MathValue<'a>

Distance between underbar and (ink) bottom of the base.

pub fn underbar_rule_thickness(&self) -> MathValue<'a>

Thickness of underbar.

pub fn underbar_extra_descender(&self) -> MathValue<'a>

Extra white space reserved below the underbar.

pub fn radical_vertical_gap(&self) -> MathValue<'a>

Space between the (ink) top of the expression and the bar over it.

pub fn radical_display_style_vertical_gap(&self) -> MathValue<'a>

Space between the (ink) top of the expression and the bar over it.

pub fn radical_rule_thickness(&self) -> MathValue<'a>

Thickness of the radical rule.

pub fn radical_extra_ascender(&self) -> MathValue<'a>

Extra white space reserved above the radical.

pub fn radical_kern_before_degree(&self) -> MathValue<'a>

Extra horizontal kern before the degree of a radical, if such is present.

pub fn radical_kern_after_degree(&self) -> MathValue<'a>

Negative kern after the degree of a radical, if such is present.

pub fn radical_degree_bottom_raise_percent(&self) -> i16

Height of the bottom of the radical degree, if such is present, in proportion to the ascender of the radical sign.