ttf_parser

Module morx

An Extended Glyph Metamorphosis Table implementation.

Structs

Chain

A metamorphosis chain.

Chains

A list of metamorphosis chains.

ChainsIter

An iterator over metamorphosis chains.

ContextualEntryData

A contextual subtable state table trailing data.

ContextualSubtable

A contextual subtable.

Coverage

A subtable coverage.

Feature

The feature table is used to compute the sub-feature flags for a list of requested features and settings.

InsertionEntryData

A contextual subtable state table trailing data.

InsertionSubtable

An insertion subtable.

LigatureSubtable

A ligature subtable.

Subtable

A subtable in a metamorphosis chain.

Subtables

A list of subtables in a metamorphosis chain.

SubtablesIter

An iterator over a metamorphosis chain subtables.

Table

An Extended Glyph Metamorphosis Table.

Enums

SubtableKind

A subtable kind.