ttf_parser::parser

Struct Stream

pub struct Stream<'a> {
    data: &'a [u8],
    offset: usize,
}

A streaming binary parser.

Fields
data: &'a [u8]offset: usize
Implementations

impl<'a> Stream<'a>

pub fn new(data: &'a [u8]) -> Self

Creates a new Stream parser.

pub fn new_at(data: &'a [u8], offset: usize) -> Option<Self>

Creates a new Stream parser at offset.

Returns None when offset is out of bounds.

pub fn at_end(&self) -> bool

Checks that stream reached the end of the data.

pub fn jump_to_end(&mut self)

Jumps to the end of the stream.

Useful to indicate that we parsed all the data.

pub fn offset(&self) -> usize

Returns the current offset.

pub fn tail(&self) -> Option<&'a [u8]>

Returns the trailing data.

Returns None when Stream is reached the end.

pub fn skip<T: FromData>(&mut self)

Advances by FromData::SIZE.

Doesnโt check bounds.

pub fn advance(&mut self, len: usize)

Advances by the specified len.

Doesnโt check bounds.

pub fn advance_checked(&mut self, len: usize) -> Option<()>

Advances by the specified len and checks for bounds.

pub fn read<T: FromData>(&mut self) -> Option<T>

Parses the type from the steam.

Returns None when there is not enough data left in the stream or the type parsing failed.

pub fn read_at<T: FromData>(data: &[u8], offset: usize) -> Option<T>

Parses the type from the steam at offset.

pub fn read_bytes(&mut self, len: usize) -> Option<&'a [u8]>

Reads N bytes from the stream.

pub fn read_array16<T: FromData>( &mut self, count: u16, ) -> Option<LazyArray16<'a, T>>

Reads the next count types as a slice.

pub fn read_array32<T: FromData>( &mut self, count: u32, ) -> Option<LazyArray32<'a, T>>

Reads the next count types as a slice.

pub fn read_at_offset16(&mut self, data: &'a [u8]) -> Option<&'a [u8]>