Struct ttf_parser::FaceTables

pub struct FaceTables<'a> { /* private fields */ }

Parsed face tables.

This struct adds the from_table_provider() method that is not available on the Font. You can create a FaceTables struct from your own, custom font provider. This is important if your font provider does things that ttf-parser currently doesn’t implement (for example zlib / brotli decoding)

Implementations

impl<'a> FaceTables<'a>

pub fn from_table_provider<T>(provider: T) -> Result<Self, FaceParsingError>where T: Iterator<Item = Result<(Tag, Option<&'a [u8]>), FaceParsingError>>,

Creates and parses face tables from an existing table provider.

This is useful for integrating ttf-parser with other font-parsing libraries that already do table decoding

pub fn has_table(&self, name: TableName) -> bool

Checks that face has a specified table.

Will return true only for tables that were successfully parsed.

pub fn names(&self) -> Names<'_>

Returns an iterator over Name Records.

An iterator can be empty.

pub fn is_regular(&self) -> bool

Checks that face is marked as Regular.

Returns false when OS/2 table is not present.

pub fn is_italic(&self) -> bool

Checks that face is marked as Italic.

Returns false when OS/2 table is not present.

pub fn is_bold(&self) -> bool

Checks that face is marked as Bold.

Returns false when OS/2 table is not present.

pub fn is_oblique(&self) -> bool

Checks that face is marked as Oblique.

Returns false when OS/2 table is not present or when its version is < 4.

pub fn is_monospaced(&self) -> bool

Checks that face is marked as Monospaced.

Returns false when post table is not present.

pub fn is_variable(&self) -> bool

Checks that face is variable.

Simply checks the presence of a fvar table.

pub fn weight(&self) -> Weight

Returns face’s weight.

Returns Weight::Normal when OS/2 table is not present.

pub fn width(&self) -> Width

Returns face’s width.

Returns Width::Normal when OS/2 table is not present or when value is invalid.

pub fn italic_angle(&self) -> Option<f32>

Returns face’s italic angle.

Returns None when post table is not present.

pub fn ascender(&self) -> i16

Returns a horizontal face ascender.

This method is affected by variation axes.

pub fn descender(&self) -> i16

Returns a horizontal face descender.

This method is affected by variation axes.

pub fn height(&self) -> i16

Returns face’s height.

This method is affected by variation axes.

pub fn line_gap(&self) -> i16

Returns a horizontal face line gap.

This method is affected by variation axes.

pub fn typographic_ascender(&self) -> Option<i16>

Returns a horizontal typographic face ascender.

Prefer Face::ascender unless you explicitly want this. This is a more low-level alternative.

This method is affected by variation axes.

Returns None when OS/2 table is not present.

pub fn typographic_descender(&self) -> Option<i16>

Returns a horizontal typographic face descender.

Prefer Face::descender unless you explicitly want this. This is a more low-level alternative.

This method is affected by variation axes.

Returns None when OS/2 table is not present.

pub fn typographic_line_gap(&self) -> Option<i16>

Returns a horizontal typographic face line gap.

Prefer Face::line_gap unless you explicitly want this. This is a more low-level alternative.

This method is affected by variation axes.

Returns None when OS/2 table is not present.

pub fn vertical_ascender(&self) -> Option<i16>

Returns a vertical face ascender.

This method is affected by variation axes.

pub fn vertical_descender(&self) -> Option<i16>

Returns a vertical face descender.

This method is affected by variation axes.

pub fn vertical_height(&self) -> Option<i16>

Returns a vertical face height.

This method is affected by variation axes.

pub fn vertical_line_gap(&self) -> Option<i16>

Returns a vertical face line gap.

This method is affected by variation axes.

pub fn units_per_em(&self) -> Option<u16>

Returns face’s units per EM.

Returns None when value is not in a 16..=16384 range.

pub fn x_height(&self) -> Option<i16>

Returns face’s x height.

This method is affected by variation axes.

Returns None when OS/2 table is not present or when its version is < 2.

pub fn capital_height(&self) -> Option<i16>

Returns face’s capital height.

This method is affected by variation axes.

Returns None when OS/2 table is not present or when its version is < 2.

pub fn underline_metrics(&self) -> Option<LineMetrics>

Returns face’s underline metrics.

This method is affected by variation axes.

Returns None when post table is not present.

pub fn strikeout_metrics(&self) -> Option<LineMetrics>

Returns face’s strikeout metrics.

This method is affected by variation axes.

Returns None when OS/2 table is not present.

pub fn subscript_metrics(&self) -> Option<ScriptMetrics>

Returns face’s subscript metrics.

This method is affected by variation axes.

Returns None when OS/2 table is not present.

pub fn superscript_metrics(&self) -> Option<ScriptMetrics>

Returns face’s superscript metrics.

This method is affected by variation axes.

Returns None when OS/2 table is not present.

pub fn number_of_glyphs(&self) -> u16

Returns a total number of glyphs in the face.

Never zero.

The value was already parsed, so this function doesn’t involve any parsing.

pub fn character_mapping_subtables(&self) -> Subtables<'_>

Returns an iterator over character to glyph index mapping.

This is a more low-level alternative to Face::glyph_index.

An iterator can be empty.

pub fn glyph_index(&self, c: char) -> Option<GlyphId>

Resolves a Glyph ID for a code point.

Returns None instead of 0 when glyph is not found.

All subtable formats except Mixed Coverage (8) are supported.

If you need a more low-level control, prefer Face::character_mapping_subtables.

pub fn glyph_variation_index(&self, c: char, variation: char) -> Option<GlyphId>

Resolves a variation of a Glyph ID from two code points.

Implemented according to Unicode Variation Sequences.

Returns None instead of 0 when glyph is not found.

pub fn glyph_hor_advance(&self, glyph_id: GlyphId) -> Option<u16>

Returns glyph’s horizontal advance.

This method is affected by variation axes.

pub fn glyph_ver_advance(&self, glyph_id: GlyphId) -> Option<u16>

Returns glyph’s vertical advance.

This method is affected by variation axes.

pub fn glyph_hor_side_bearing(&self, glyph_id: GlyphId) -> Option<i16>

Returns glyph’s horizontal side bearing.

This method is affected by variation axes.

pub fn glyph_ver_side_bearing(&self, glyph_id: GlyphId) -> Option<i16>

Returns glyph’s vertical side bearing.

This method is affected by variation axes.

pub fn glyph_y_origin(&self, glyph_id: GlyphId) -> Option<i16>

Returns glyph’s vertical origin according to Vertical Origin Table.

pub fn glyph_name(&self, glyph_id: GlyphId) -> Option<&str>

Returns glyph’s name.

Uses the post and CFF tables as sources.

Returns None when no name is associated with a glyph.

pub fn has_glyph_classes(&self) -> bool

Checks that face has Glyph Class Definition Table.

pub fn glyph_class(&self, glyph_id: GlyphId) -> Option<GlyphClass>

Returns glyph’s class according to Glyph Class Definition Table.

Returns None when Glyph Class Definition Table is not set or glyph class is not set or invalid.

pub fn glyph_mark_attachment_class(&self, glyph_id: GlyphId) -> Class

Returns glyph’s mark attachment class according to Mark Attachment Class Definition Table.

All glyphs not assigned to a class fall into Class 0.

pub fn is_mark_glyph(&self, glyph_id: GlyphId, set_index: Option<u16>) -> bool

Checks that glyph is a mark according to Mark Glyph Sets Table.

set_index allows checking a specific glyph coverage set. Otherwise all sets will be checked.

pub fn kerning_subtables(&self) -> Subtables<'_>

Returns a iterator over kerning subtables.

Supports both OpenType and Apple Advanced Typography variants.

pub fn outline_glyph( &self, glyph_id: GlyphId, builder: &mut dyn OutlineBuilder ) -> Option<Rect>

Outlines a glyph and returns its tight bounding box.

Warning: since ttf-parser is a pull parser, OutlineBuilder will emit segments even when outline is partially malformed. You must check outline_glyph() result before using OutlineBuilder’s output.

gvar, glyf, CFF and CFF2 tables are supported. And they will be accesses in this specific order.

This method is affected by variation axes.

Returns None when glyph has no outline or on error.

Example

use std::fmt::Write;
use ttf_parser;

struct Builder(String);

impl ttf_parser::OutlineBuilder for Builder {
    fn move_to(&mut self, x: f32, y: f32) {
        write!(&mut self.0, "M {} {} ", x, y).unwrap();
    }

    fn line_to(&mut self, x: f32, y: f32) {
        write!(&mut self.0, "L {} {} ", x, y).unwrap();
    }

    fn quad_to(&mut self, x1: f32, y1: f32, x: f32, y: f32) {
        write!(&mut self.0, "Q {} {} {} {} ", x1, y1, x, y).unwrap();
    }

    fn curve_to(&mut self, x1: f32, y1: f32, x2: f32, y2: f32, x: f32, y: f32) {
        write!(&mut self.0, "C {} {} {} {} {} {} ", x1, y1, x2, y2, x, y).unwrap();
    }

    fn close(&mut self) {
        write!(&mut self.0, "Z ").unwrap();
    }
}

let data = std::fs::read("tests/fonts/demo.ttf").unwrap();
let face = ttf_parser::Face::from_slice(&data, 0).unwrap();
let mut builder = Builder(String::new());
let bbox = face.outline_glyph(ttf_parser::GlyphId(1), &mut builder).unwrap();
assert_eq!(builder.0, "M 173 267 L 369 267 L 270 587 L 173 267 Z M 6 0 L 224 656 \
                       L 320 656 L 541 0 L 452 0 L 390 200 L 151 200 L 85 0 L 6 0 Z ");
assert_eq!(bbox, ttf_parser::Rect { x_min: 6, y_min: 0, x_max: 541, y_max: 656 });

pub fn glyph_bounding_box(&self, glyph_id: GlyphId) -> Option<Rect>

Returns a tight glyph bounding box.

This is just a shorthand for outline_glyph() since only the glyf table stores a bounding box. We ignore glyf table bboxes because they can be malformed. In case of CFF and variable fonts we have to actually outline a glyph to find it’s bounding box.

When a glyph is defined by a raster or a vector image, that can be obtained via glyph_image(), the bounding box must be calculated manually and this method will return None.

Note: the returned bbox is not validated in any way. A font file can have a glyph bbox set to zero/negative width and/or height and this is perfectly ok. For calculated bboxes, zero width and/or height is also perfectly fine.

This method is affected by variation axes.

pub fn global_bounding_box(&self) -> Rect

Returns a bounding box that large enough to enclose any glyph from the face.

pub fn glyph_raster_image( &self, glyph_id: GlyphId, pixels_per_em: u16 ) -> Option<RasterGlyphImage<'_>>

Returns a reference to a glyph’s raster image.

A font can define a glyph using a raster or a vector image instead of a simple outline. Which is primarily used for emojis. This method should be used to access raster images.

pixels_per_em allows selecting a preferred image size. The chosen size will be closer to an upper one. So when font has 64px and 96px images and pixels_per_em is set to 72, 96px image will be returned. To get the largest image simply use std::u16::MAX.

Note that this method will return an encoded image. It should be decoded by the caller. We don’t validate or preprocess it in any way.

Currently, only PNG images are supported.

Also, a font can contain both: images and outlines. So when this method returns None you should also try outline_glyph() afterwards.

There are multiple ways an image can be stored in a TrueType font and this method supports only sbix, CBLC+CBDT. Font’s tables be accesses in this specific order.

pub fn glyph_svg_image(&self, glyph_id: GlyphId) -> Option<&'a [u8]>

Returns a reference to a glyph’s SVG image.

A font can define a glyph using a raster or a vector image instead of a simple outline. Which is primarily used for emojis. This method should be used to access SVG images.

Note that this method will return just an SVG data. It should be rendered or even decompressed (in case of SVGZ) by the caller. We don’t validate or preprocess it in any way.

Also, a font can contain both: images and outlines. So when this method returns None you should also try outline_glyph() afterwards.