Struct ttf_parser::Font

pub struct Font<'a> {

    avar: Option<Table<'a>>,
    cbdt: Option<&'a [u8]>,
    cblc: Option<&'a [u8]>,
    cff_: Option<Metadata<'a>>,
    cff2: Option<Metadata<'a>>,
    cmap: Option<Table<'a>>,
    fvar: Option<Table<'a>>,
    gdef: Option<Table<'a>>,
    glyf: Option<&'a [u8]>,
    gvar: Option<Table<'a>>,
    head: &'a [u8],
    hhea: &'a [u8],
    hmtx: Option<Table<'a>>,
    hvar: Option<Table<'a>>,
    kern: Option<Subtables<'a>>,
    loca: Option<Table<'a>>,
    mvar: Option<Table<'a>>,
    name: Option<Names<'a>>,
    os_2: Option<Table<'a>>,
    post: Option<Table<'a>>,
    vhea: Option<&'a [u8]>,
    vmtx: Option<Table<'a>>,
    sbix: Option<&'a [u8]>,
    svg_: Option<&'a [u8]>,
    vorg: Option<Table<'a>>,
    vvar: Option<Table<'a>>,
    number_of_glyphs: NonZeroU16,
    coordinates: VarCoords,

}

A font data handle.

Fields
avar: Option<Table<'a>>cbdt: Option<&'a [u8]>cblc: Option<&'a [u8]>cff_: Option<Metadata<'a>>cff2: Option<Metadata<'a>>cmap: Option<Table<'a>>fvar: Option<Table<'a>>gdef: Option<Table<'a>>glyf: Option<&'a [u8]>gvar: Option<Table<'a>>head: &'a [u8]hhea: &'a [u8]hmtx: Option<Table<'a>>hvar: Option<Table<'a>>kern: Option<Subtables<'a>>loca: Option<Table<'a>>mvar: Option<Table<'a>>name: Option<Names<'a>>os_2: Option<Table<'a>>post: Option<Table<'a>>vhea: Option<&'a [u8]>vmtx: Option<Table<'a>>sbix: Option<&'a [u8]>svg_: Option<&'a [u8]>vorg: Option<Table<'a>>vvar: Option<Table<'a>>number_of_glyphs: NonZeroU16coordinates: VarCoords
Implementations

impl<'a> Font<'a>

pub fn from_data(data: &'a [u8], index: u32) -> Option<Self>

Creates a Font object from a raw data.

You can set index for font collections. For simple ttf fonts set index to 0.

This method will do some parsing and sanitization, so it’s a bit expensive.

Required tables: head, hhea and maxp.

If an optional table has an invalid data it will be skipped.

pub fn has_table(&self, name: TableName) -> bool

Checks that font has a specified table.

Will return true only for tables that were successfully parsed.

pub fn names(&self) -> Names<'_>

Returns an iterator over Name Records.

An iterator can be empty.

pub fn family_name(&self) -> Option<String>

Returns font’s family name.

Typographic Family is preferred over Family.

Note that font can have multiple names. You can use names() to list them all.

pub fn post_script_name(&self) -> Option<String>

Returns font’s PostScript name.

Note that font can have multiple names. You can use names() to list them all.

pub fn is_regular(&self) -> bool

Checks that font is marked as Regular.

Returns false when OS/2 table is not present.

pub fn is_italic(&self) -> bool

Checks that font is marked as Italic.

Returns false when OS/2 table is not present.

pub fn is_bold(&self) -> bool

Checks that font is marked as Bold.

Returns false when OS/2 table is not present.

pub fn is_oblique(&self) -> bool

Checks that font is marked as Oblique.

Returns false when OS/2 table is not present or when its version is < 4.

pub fn is_variable(&self) -> bool

Checks that font is variable.

Simply checks the presence of a fvar table.

pub fn weight(&self) -> Weight

Returns font’s weight.

Returns Weight::Normal when OS/2 table is not present.

pub fn width(&self) -> Width

Returns font’s width.

Returns Width::Normal when OS/2 table is not present or when value is invalid.

pub(crate) fn use_typo_metrics(&self) -> Option<Table<'_>>

pub fn ascender(&self) -> i16

Returns a horizontal font ascender.

This method is affected by variation axes.

pub fn descender(&self) -> i16

Returns a horizontal font descender.

This method is affected by variation axes.

pub fn height(&self) -> i16

Returns font’s height.

This method is affected by variation axes.

pub fn line_gap(&self) -> i16

Returns a horizontal font line gap.

This method is affected by variation axes.

pub fn vertical_ascender(&self) -> Option<i16>

Returns a vertical font ascender.

This method is affected by variation axes.

pub fn vertical_descender(&self) -> Option<i16>

Returns a vertical font descender.

This method is affected by variation axes.

pub fn vertical_height(&self) -> Option<i16>

Returns a vertical font height.

This method is affected by variation axes.

pub fn vertical_line_gap(&self) -> Option<i16>

Returns a vertical font line gap.

This method is affected by variation axes.

pub fn units_per_em(&self) -> Option<u16>

Returns font’s units per EM.

Returns None when value is not in a 16..=16384 range.

pub fn x_height(&self) -> Option<i16>

Returns font’s x height.

This method is affected by variation axes.

Returns None when OS/2 table is not present or when its version is < 2.

pub fn underline_metrics(&self) -> Option<LineMetrics>

Returns font’s underline metrics.

This method is affected by variation axes.

Returns None when post table is not present.

pub fn strikeout_metrics(&self) -> Option<LineMetrics>

Returns font’s strikeout metrics.

This method is affected by variation axes.

Returns None when OS/2 table is not present.

pub fn subscript_metrics(&self) -> Option<ScriptMetrics>

Returns font’s subscript metrics.

This method is affected by variation axes.

Returns None when OS/2 table is not present.

pub fn superscript_metrics(&self) -> Option<ScriptMetrics>

Returns font’s superscript metrics.

This method is affected by variation axes.

Returns None when OS/2 table is not present.

pub fn number_of_glyphs(&self) -> u16

Returns a total number of glyphs in the font.

Never zero.

The value was already parsed, so this function doesn’t involve any parsing.

pub fn glyph_index(&self, c: char) -> Option<GlyphId>

Resolves a Glyph ID for a code point.

Returns None instead of 0 when glyph is not found.

All subtable formats except Mixed Coverage (8) are supported.

pub fn glyph_variation_index(&self, c: char, variation: char) -> Option<GlyphId>

Resolves a variation of a Glyph ID from two code points.

Implemented according to Unicode Variation Sequences.

Returns None instead of 0 when glyph is not found.

pub fn glyph_hor_advance(&self, glyph_id: GlyphId) -> Option<u16>

Returns glyph’s horizontal advance.

This method is affected by variation axes.

pub fn glyph_ver_advance(&self, glyph_id: GlyphId) -> Option<u16>

Returns glyph’s vertical advance.

This method is affected by variation axes.

pub fn glyph_hor_side_bearing(&self, glyph_id: GlyphId) -> Option<i16>

Returns glyph’s horizontal side bearing.

This method is affected by variation axes.

pub fn glyph_ver_side_bearing(&self, glyph_id: GlyphId) -> Option<i16>

Returns glyph’s vertical side bearing.

This method is affected by variation axes.

pub fn glyph_y_origin(&self, glyph_id: GlyphId) -> Option<i16>

Returns glyph’s vertical origin according to Vertical Origin Table.

pub fn glyph_name(&self, glyph_id: GlyphId) -> Option<&str>

Returns glyph’s name.

Uses the post table as a source.

Returns None when no name is associated with a glyph.

pub fn has_glyph_classes(&self) -> bool

Checks that font has Glyph Class Definition Table.

pub fn glyph_class(&self, glyph_id: GlyphId) -> Option<GlyphClass>

Returns glyph’s class according to Glyph Class Definition Table.

Returns None when Glyph Class Definition Table is not set or glyph class is not set or invalid.

pub fn glyph_mark_attachment_class(&self, glyph_id: GlyphId) -> Class

Returns glyph’s mark attachment class according to Mark Attachment Class Definition Table.

All glyphs not assigned to a class fall into Class 0.

pub fn is_mark_glyph(&self, glyph_id: GlyphId, set_index: Option<u16>) -> bool

Checks that glyph is a mark according to Mark Glyph Sets Table.

set_index allows checking a specific glyph coverage set. Otherwise all sets will be checked.

Returns Ok(false) when Mark Glyph Sets Table is not set.

pub fn kerning_subtables(&self) -> Subtables<'_>

Returns a iterator over kerning subtables.

Supports both OpenType and Apple Advanced Typography variants.

pub fn outline_glyph(
&self,
glyph_id: GlyphId,
builder: &mut dyn OutlineBuilder
) -> Option<Rect>

Outlines a glyph and returns its tight bounding box.

Warning: since ttf-parser is a pull parser, OutlineBuilder will emit segments even when outline is partially malformed. You must check outline_glyph() result before using OutlineBuilder’s output.

glyf, gvar, CFF and CFF2 tables are supported.

This method is affected by variation axes.

Returns None when glyph has no outline or on error.

Example

use std::fmt::Write;
use ttf_parser;

struct Builder(String);

impl ttf_parser::OutlineBuilder for Builder {
    fn move_to(&mut self, x: f32, y: f32) {
        write!(&mut self.0, "M {} {} ", x, y).unwrap();
    }

    fn line_to(&mut self, x: f32, y: f32) {
        write!(&mut self.0, "L {} {} ", x, y).unwrap();
    }

    fn quad_to(&mut self, x1: f32, y1: f32, x: f32, y: f32) {
        write!(&mut self.0, "Q {} {} {} {} ", x1, y1, x, y).unwrap();
    }

    fn curve_to(&mut self, x1: f32, y1: f32, x2: f32, y2: f32, x: f32, y: f32) {
        write!(&mut self.0, "C {} {} {} {} {} {} ", x1, y1, x2, y2, x, y).unwrap();
    }

    fn close(&mut self) {
        write!(&mut self.0, "Z ").unwrap();
    }
}

let data = std::fs::read("fonts/SourceSansPro-Regular-Tiny.ttf").unwrap();
let font = ttf_parser::Font::from_data(&data, 0).unwrap();
let mut builder = Builder(String::new());
let bbox = font.outline_glyph(ttf_parser::GlyphId(13), &mut builder).unwrap();
assert_eq!(builder.0, "M 90 0 L 90 656 L 173 656 L 173 71 L 460 71 L 460 0 L 90 0 Z ");
assert_eq!(bbox, ttf_parser::Rect { x_min: 90, y_min: 0, x_max: 460, y_max: 656 });

pub fn glyph_bounding_box(&self, glyph_id: GlyphId) -> Option<Rect>

Returns a tight glyph bounding box.

Unless the current font has a glyf table, this is just a shorthand for outline_glyph() since only the glyf table stores a bounding box. In case of CFF and variable fonts we have to actually outline a glyph to find it’s bounding box.

When a glyph is defined by a raster or a vector image, that can be obtained via glyph_image(), the bounding box must be calculated manually and this method will return None.

This method is affected by variation axes.

pub fn glyph_raster_image(
&self,
glyph_id: GlyphId,
pixels_per_em: u16
) -> Option<RasterGlyphImage<'_>>

Returns a reference to a glyph’s raster image.

A font can define a glyph using a raster or a vector image instead of a simple outline. Which is primarily used for emojis. This method should be used to access raster images.

pixels_per_em allows selecting a preferred image size. The chosen size will be closer to an upper one. So when font has 64px and 96px images and pixels_per_em is set to 72, 96px image will be returned. To get the largest image simply use std::u16::MAX.

Note that this method will return an encoded image. It should be decoded by the caller. We don’t validate or preprocess it in any way.

Currently, only PNG images are supported.

Also, a font can contain both: images and outlines. So when this method returns None you should also try outline_glyph() afterwards.

There are multiple ways an image can be stored in a TrueType font and this method supports only sbix, CBLC+CBDT.

pub fn glyph_svg_image(&self, glyph_id: GlyphId) -> Option<&'a [u8]>

Returns a reference to a glyph’s SVG image.

A font can define a glyph using a raster or a vector image instead of a simple outline. Which is primarily used for emojis. This method should be used to access SVG images.

Note that this method will return just an SVG data. It should be rendered or even decompressed (in case of SVGZ) by the caller. We don’t validate or preprocess it in any way.

Also, a font can contain both: images and outlines. So when this method returns None you should also try outline_glyph() afterwards.

pub fn variation_axes(&self) -> VariationAxes<'_>

Returns an iterator over variation axes.

pub fn set_variation(&mut self, axis: Tag, value: f32) -> Option<()>

Sets a variation axis coordinate.

This is the only mutable method in the library. We can simplify the API a lot by storing the variable coordinates in the font object itself.

Since coordinates are stored on the stack, we allow only 32 of them.

Returns None when font is not variable or doesn’t have such axis.

pub(crate) fn metrics_var_offset(&self, tag: Tag) -> f32

pub(crate) fn apply_metrics_variation(&self, tag: Tag, value: i16) -> i16

pub(crate) fn apply_metrics_variation_to(&self, tag: Tag, value: &mut i16)

pub(crate) fn coords(&self) -> &[NormalizedCoord]