ttf_parser

Struct LazyArray32

pub struct LazyArray32<'a, T> { /* private fields */ }

A slice-like container that converts internal binary data only on access.

This is a low-level, internal structure that should not be used directly.

Implementations

impl<'a, T: FromData> LazyArray32<'a, T>

pub fn new(data: &'a [u8]) -> Self

Creates a new LazyArray.

pub fn get(&self, index: u32) -> Option<T>

Returns a value at index.

pub fn len(&self) -> u32

Returns arrayโs length.

pub fn is_empty(&self) -> bool

Checks if the array is empty.

pub fn binary_search(&self, key: &T) -> Option<(u32, T)>
where T: Ord,

Performs a binary search by specified key.

pub fn binary_search_by<F>(&self, f: F) -> Option<(u32, T)>
where F: FnMut(&T) -> Ordering,

Performs a binary search using specified closure.