Module ttf_parser::tables::cff

Modules

operator

Enumerates some operators defined in the Adobe Technical Note #5177.

private_dict_operator

Enumerates some operators defined in the Adobe Technical Note #5176, Table 23 Private DICT Operators

top_dict_operator

Enumerates some operators defined in the Adobe Technical Note #5176, Table 9 Top DICT Operator Entries

Structs

ArgumentsStack

Builder

CharStringParserContext

DataIndex

DataIndexIter

DictionaryParser

Metadata

Operator

VarOffsets

Enums

CFFError

A list of errors that can occur during a CFF table parsing.

OffsetSize

Constants

END_OF_FLOAT_FLAG

MAX_ARGUMENTS_STACK_LEN

MAX_OPERANDS_LEN

STACK_LIMIT

TWO_BYTE_OPERATOR_MARK

Traits

IsEven

Functions

_parse_char_string

calc_subroutine_bias

conv_subroutine_index

f32_abs

is_dict_one_byte_op

outline

parse_char_string

parse_index

parse_index_impl

parse_metadata

parse_number

parse_private_dict

parse_top_dict

skip_index

skip_number

try_parse_offset_size