ttf_parser::tables::cff::dict

Struct DictionaryParser

pub struct DictionaryParser<'a> {
    data: &'a [u8],
    offset: usize,
    operands_offset: usize,
    operands: &'a mut [f64],
    operands_len: u16,
}

Fields
data: &'a [u8]offset: usizeoperands_offset: usizeoperands: &'a mut [f64]operands_len: u16
Implementations

impl<'a> DictionaryParser<'a>

pub fn new(data: &'a [u8], operands_buffer: &'a mut [f64]) -> Self

pub fn parse_next(&mut self) -> Option<Operator>

pub fn parse_operands(&mut self) -> Option<()>

Parses operands of the current operator.

In the DICT structure, operands are defined before an operator. So we are trying to find an operator first and the we can actually parse the operands.

Since this methods is pretty expensive and we do not care about most of the operators, we can speed up parsing by parsing operands only for required operators.

We still have to “skip” operands during operators search (see skip_number()), but it’s still faster that a naive method.

pub fn operands(&self) -> &[f64]

pub fn parse_number(&mut self) -> Option<f64>

pub fn parse_offset(&mut self) -> Option<usize>

pub fn parse_range(&mut self) -> Option<Range<usize>>