Crate ttf_parser

A high-level, safe, zero-allocation font parser for:

TrueType,

OpenType, and

AAT fonts.

Font parsing starts with a Face.

Features

A high-level API for most common properties, hiding all parsing and data resolving logic.

A low-level, but safe API to access TrueType tables data.

Highly configurable. You can disable most of the features, reducing binary size. You can also parse TrueType tables separately, without loading the whole font/face.

Zero heap allocations.

Zero unsafe.

Zero dependencies.

no_std/WASM compatible.

Fast.

Stateless. All parsing methods are immutable.

Simple and maintainable code (no magic numbers).

Safety

The library must not panic. Any panic considered as a critical bug and should be reported.

The library forbids unsafe code.

No heap allocations, so crash due to OOM is not possible.

All recursive methods have a depth limit.

Technically, should use less than 64KiB of stack in worst case scenario.

Most of arithmetic operations are checked.

Most of numeric casts are checked.

Modules

ankr

An Anchor Point Table implementation.

apple_layout

This module contains Apple Advanced Typography Layout supplementary tables implementation.

avar

An Axis Variations Table implementation.

cbdt

A Color Bitmap Data Table implementation.

cblc

A Color Bitmap Location Table implementation.

cff

A Compact Font Format Table implementation.

cff2

A Compact Font Format 2 Table implementation.

cmap

A Character to Glyph Index Mapping Table implementation.

colr

A Color Table implementation.

cpal

A Color Palette Table implementation.

feat

A Feature Name Table implementation.

fvar

A Font Variations Table implementation.

gdef

A Glyph Definition Table implementation.

glyf

A Glyph Data Table implementation.

gpos

A Glyph Positioning Table implementation.

gsub

A Glyph Substitution Table implementation.

gvar

A Glyph Variations Table implementation.

head

A Font Header Table implementation.

hhea

A Horizontal Header Table implementation.

hmtx

A Horizontal/Vertical Metrics Table implementation.

hvar

A Horizontal Metrics Variations Table implementation.

kern

A Kerning Table implementation.

kerx

An Extended Kerning Table implementation.

loca

An Index to Location Table implementation.

math

A Math Table implementation.

maxp

A Maximum Profile Table implementation.

morx

An Extended Glyph Metamorphosis Table implementation.

mvar

A Metrics Variations Table implementation.

name

A Naming Table implementation.

name_id

A list of name ID’s.

opentype_layout

This module contains OpenType Layout supplementary tables implementation.

os2

A OS/2 and Windows Metrics Table implementation.

post

A PostScript Table implementation.

sbix

A Standard Bitmap Graphics Table implementation.

stat

A Style Attributes Table implementation.

svg

An SVG Table implementation.

trak

A Tracking Table implementation.

vhea

A Vertical Header Table implementation.

vorg

A Vertical Origin Table implementation.

vvar

A Vertical Metrics Variations Table implementation.

Structs

Face

A font face.

FaceTables

Parsed face tables.

Fixed

A 32-bit signed fixed-point number (16.16).

GlyphId

A type-safe wrapper for glyph ID.

LazyArray16

A slice-like container that converts internal binary data only on access.

LazyArray32

A slice-like container that converts internal binary data only on access.

LazyArrayIter16

An iterator over LazyArray16.

LazyArrayIter32

An iterator over LazyArray32.

LineMetrics

A line metrics.

NormalizedCoordinate

A variation coordinate in a normalized coordinate system.

PhantomPoints

Phantom points.

PointF

A float point.

RasterGlyphImage

A glyph’s raster image.

RawFace

A raw font face.

RawFaceTables

A list of all supported tables as raw data.

Rect

A rectangle.

RectF

A rectangle described by the left-lower and upper-right points.

RgbaColor

A RGBA color in the sRGB color space.

ScriptMetrics

A script metrics used by subscript and superscript.

TableRecord

A raw table record.

Tag

A 4-byte tag.

Transform

An affine transform.

UnicodeRanges

Unicode Ranges.

Variation

A font variation value.

VariationAxis

A variation axis.

Enums

CFFError

A list of errors that can occur during a CFF glyph outlining.

FaceParsingError

A list of font face parsing errors.

Language

A Name language.

Permissions

Face permissions.

PlatformId

A platform ID.

RasterImageFormat

A glyph raster image format.

Style

A face style.

Weight

A face weight.

Width

A face width.

Traits

FromData

A trait for parsing raw binary data of fixed size.

OutlineBuilder

A trait for glyph outline construction.

Functions

fonts_in_collection

Returns the number of fonts stored in a TrueType font collection.