Trait ttf_parser::colr::Painter

pub trait Painter<'a> {

    // Required methods
    fn outline_glyph(&mut self, glyph_id: GlyphId);
    fn paint(&mut self, paint: Paint<'a>);
    fn push_clip(&mut self);
    fn push_clip_box(&mut self, clipbox: ClipBox);
    fn pop_clip(&mut self);
    fn push_layer(&mut self, mode: CompositeMode);
    fn pop_layer(&mut self);
    fn push_translate(&mut self, tx: f32, ty: f32);
    fn push_scale(&mut self, sx: f32, sy: f32);
    fn push_rotate(&mut self, angle: f32);
    fn push_skew(&mut self, skew_x: f32, skew_y: f32);
    fn push_transform(&mut self, transform: Transform);
    fn pop_transform(&mut self);

}

A trait for color glyph painting.

See COLR for details.

Required Methods

fn outline_glyph(&mut self, glyph_id: GlyphId)

Outline a glyph and store it.

fn paint(&mut self, paint: Paint<'a>)

Paint the stored outline using the provided color.

fn push_clip(&mut self)

Push a new clip path using the currently stored outline.

fn push_clip_box(&mut self, clipbox: ClipBox)

Push a new clip path using the clip box.

fn pop_clip(&mut self)

Pop the last clip path.

fn push_layer(&mut self, mode: CompositeMode)

Push a new layer with the given composite mode.

fn pop_layer(&mut self)

Pop the last layer.

fn push_translate(&mut self, tx: f32, ty: f32)

Push a translation transform.

fn push_scale(&mut self, sx: f32, sy: f32)

Push a scaling transform.

fn push_rotate(&mut self, angle: f32)

Push a rotation transform.

fn push_skew(&mut self, skew_x: f32, skew_y: f32)

Push a skewing transform.

fn push_transform(&mut self, transform: Transform)

Push a transform.

fn pop_transform(&mut self)

Pop the last transform.