Module ttf_parser::parser

Binary parsing utils.

This module should not be used directly, unless you’re planning to parse some tables manually.

Structs

F2DOT14
A 16-bit signed fixed number with the low 14 bits of fraction (2.14).

Fixed
A 32-bit signed fixed-point number (16.16).

LazyArray16
A slice-like container that converts internal binary data only on access.

LazyArray32
A slice-like container that converts internal binary data only on access.

LazyArrayIter16
An iterator over LazyArray16.

LazyArrayIter32
An iterator over LazyArray32.

Offset16
A type-safe u16 offset.

Offset32
A type-safe u32 offset.

Stream
A streaming binary parser.

U24
A u24 number.

Traits

FromData
A trait for parsing raw binary data.

NumFrom
A safe u32 to usize casting.

Offset
A common offset methods.

TryNumFrom
Just like TryFrom, but for numeric types not supported by the Rust’s std.