Module ttf_parser::parser

Structs

F2DOT14

A 16-bit signed fixed number with the low 14 bits of fraction (2.14).

Fixed

A 32-bit signed fixed-point number (16.16).

LazyArray16

A slice-like container that converts internal binary data only on access.

LazyArray32

A slice-like container that converts internal binary data only on access.

LazyArrayIter16

An iterator over LazyArray16.

LazyArrayIter32

An iterator over LazyArray32.

Offset16

Offset32

Stream

U24

Traits

FromData

A trait for parsing raw binary data.

NumFrom

Offset

TryNumFrom

Just like TryFrom, but for numeric types not supported by the Rust’s std.

Functions

f32_bound

i16_bound