Trait ttf_parser::parser::FromData

pub trait FromData: Sized {
    const SIZE: usize;

    // Required method
    fn parse(data: &[u8]) -> Option<Self>;
}

A trait for parsing raw binary data of fixed size.

This is a low-level, internal trait that should not be used directly.

Required Associated Constants

const SIZE: usize

Object’s raw data size.

Not always the same as mem::size_of.

Required Methods

fn parse(data: &[u8]) -> Option<Self>

Parses an object from a raw data.

Object Safety

This trait is not object safe.