ttf_parser::parser

Trait FromSlice

pub trait FromSlice<'a>: Sized {
    // Required method
    fn parse(data: &'a [u8]) -> Option<Self>;
}

A trait for parsing raw binary data of variable size.

This is a low-level, internal trait that should not be used directly.

Required Methods

fn parse(data: &'a [u8]) -> Option<Self>

Parses an object from a raw data.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.