ttf_parser
Struct Face

pub struct Face<'a> { /* private fields */ }

A font face handle.

Implementations

impl<'a> Face<'a>

pub fn from_slice(data: &'a [u8], index: u32) -> Result<Self, FaceParsingError>

Creates a new Face object from a raw data.

index indicates the specific font face in a font collection. Use fonts_in_collection to get the total number of font faces. Set to 0 if unsure.

This method will do some parsing and sanitization, so it’s a bit expensive.

Required tables: head, hhea and maxp.

If an optional table has invalid data it will be skipped.

pub fn table_data(&self, tag: Tag) -> Option<&'a [u8]>

Returns the raw data of a selected table.

Useful if you want to parse the data manually.