ttf_parser::tables::cff

Module cff1

A Compact Font Format Table implementation.

Modules

charset_id ğŸ”’

Enumerates Charset IDs defined in the Adobe Technical Note #5176, Table 22

encoding_id ğŸ”’

Enumerates Charset IDs defined in the Adobe Technical Note #5176, Table 16

operator ğŸ”’

Enumerates some operators defined in the Adobe Technical Note #5177.

private_dict_operator ğŸ”’

Enumerates some operators defined in the Adobe Technical Note #5176, Table 23 Private DICT Operators

top_dict_operator ğŸ”’

Enumerates some operators defined in the Adobe Technical Note #5176, Table 9 Top DICT Operator Entries

Structs

CIDMetadata ğŸ”’

CharStringParserContext ğŸ”’

Matrix

An affine transformation matrix.

PrivateDict ğŸ”’

SIDMetadata ğŸ”’

Table

A Compact Font Format Table.

TopDict ğŸ”’

Enums

FDSelect ğŸ”’

FontKind ğŸ”’

Constants

MAX_ARGUMENTS_STACK_LEN ğŸ”’

MAX_OPERANDS_LEN ğŸ”’

STACK_LIMIT ğŸ”’

TWO_BYTE_OPERATOR_MARK ğŸ”’

Functions

_parse_char_string ğŸ”’

parse_char_string ğŸ”’

parse_cid_local_subrs ğŸ”’

In CID fonts, to get local subroutines we have to:

parse_cid_metadata ğŸ”’

parse_fd_select ğŸ”’

parse_font_dict ğŸ”’

parse_private_dict ğŸ”’

parse_sid_metadata ğŸ”’

parse_top_dict ğŸ”’

seac_code_to_glyph_id ğŸ”’