Module ttf_parser::tables::cff::cff1

A Compact Font Format Table implementation.

Modules

charset_id 🔒

Enumerates Charset IDs defined in the Adobe Technical Note #5176, Table 22

encoding_id 🔒

Enumerates Charset IDs defined in the Adobe Technical Note #5176, Table 16

operator 🔒

Enumerates some operators defined in the Adobe Technical Note #5177.

private_dict_operator 🔒

Enumerates some operators defined in the Adobe Technical Note #5176, Table 23 Private DICT Operators

top_dict_operator 🔒

Enumerates some operators defined in the Adobe Technical Note #5176, Table 9 Top DICT Operator Entries

Structs

CIDMetadata 🔒

CharStringParserContext 🔒

Matrix

An affine transformation matrix.

PrivateDict 🔒

SIDMetadata 🔒

Table

A Compact Font Format Table.

TopDict 🔒

Enums

FDSelect 🔒

FontKind 🔒

Constants

MAX_ARGUMENTS_STACK_LEN 🔒

MAX_OPERANDS_LEN 🔒

STACK_LIMIT 🔒

TWO_BYTE_OPERATOR_MARK 🔒

Functions

_parse_char_string 🔒

parse_char_string 🔒

parse_cid_local_subrs 🔒

In CID fonts, to get local subroutines we have to:

parse_cid_metadata 🔒

parse_fd_select 🔒

parse_font_dict 🔒

parse_private_dict 🔒

parse_sid_metadata 🔒

parse_top_dict 🔒

seac_code_to_glyph_id 🔒