Module ttf_parser::tables::gvar

Modules

packed_deltas

packed_points

Structs

PointAndDelta

Table

TupleVariationHeaderData

VariationTuple

VariationTuples

A list of variation tuples.

Enums

GlyphVariationDataOffsets

Constants

MAX_TUPLES_LEN

The maximum number of variation tuples.

PHANTOM_POINTS_LEN

’The TrueType rasterizer dynamically generates ‘phantom’ points for each glyph that represent horizontal and vertical advance widths and side bearings, and the variation data within the gvar table includes data for these phantom points.’

Functions

infer_delta

infer_deltas

Infer unreferenced deltas.

outline

outline_var_impl

parse_tuple_variation_header

parse_variation_data

parse_variation_tuples